parbs BETA!

+++ go.sh
@@ -1,48 +1,69 @@
-OS Setup
+# PARBS 
 
-Boot Arch Linux Install Medium
-Connect to the internet.
+SSID=""			# Leave empty if ethernet
+LAPTOP=false		# Power Efficiency Services
 
-iwctl
-device list
-station wlan0 scan
-station wlan0 get-networks
-station wlan0 connect NETWORK_SSID
-station wlan0 show
+# =======================================================
+#
+# OS INSTALL INSTRUCTIONS
+# RUN MANUALLY
+#
+# Boot Arch Linux Install Medium
+# Connect to the internet.
+#
+# iwctl
+# device list
+# station wlan0 scan
+# station wlan0 get-networks
+# station wlan0 connect NETWORK_SSID
+# station wlan0 show
+#
+# sudo pacman -Syu archinstall networkmanager
+# archinstall
+# reboot
+#
+# =======================================================
 
-sudo pacman -Syu archinstall networkmanager
-archinstall
-reboot
+# Connect to the internet.
+if [ -n "$SSID" ]; then
+	sudo systemctl enable NetworkManager
+	sudo systemctl start NetworkManager
+	nmcli radio wifi
+	nmcli dev wifi list
+	nmcli dev wifi connect $SSID --ask
+fi
 
-DE Setup
-
-Connect to the internet.
-
-sudo systemctl enable NetworkManager
-sudo systemctl start NetworkManager
-nmcli radio wifi
-nmcli dev wifi list
-nmcli dev wifi connect NETWORK_SSID --ask
-
-sudo pacman -S –needed git base-devel bluez bluez-utils blueman xorg-server xorg-apps xorg-xinit i3-gaps i3blocks i3lock numlockx noto-fonts ttf-ubuntu-font-family ttf-dejavu ttf-freefont ttf-liberation ttf-droid ttf-roboto terminus-font rxvt-unicode ranger rofi dmenu kitty tlp tlp-rdw powertop acpi 
-install AUR manager
-
+# install AUR manager
+cd
 git clone https://aur.archlinux.org/yay.git
 cd yay
 makepkg -si
 cd
 
-Setup
+# install packages
+sudo pacman -S --needed git base-devel bluez bluez-utils blueman xorg-server xorg-apps xorg-xinit i3-gaps i3blocks i3lock numlockx noto-fonts ttf-ubuntu-font-family ttf-dejavu ttf-freefont ttf-liberation ttf-droid ttf-roboto terminus-font rxvt-unicode ranger rofi dmenu kitty tlp tlp-rdw powertop acpi 
 
 yay -S pa-applet-git
-sudo systemctl enable bluetooth
-sudo systemctl enable tlp
-sudo systemctl enable tlp-sleep
-sudo systemctl mask systemd-rfkill.service
-sudo systemctl mask systemd-rfkill.socket
-sudo systemctl enable fstrim.timer
 
+# Get dots!
+cd
+git clone https://phylopylo/dots.git
+cp dots/main/linux-bashrc ~/.bashrc
+cp dots/main/linux-tmux.conf ~/.tmux.conf
+
+# Enable Services
+sudo systemctl enable bluetooth
+sudo systemctl enable fstrim.timer
 sudo systemctl enable lightdm
+
+# Power Efficiency Settings
+if [ "$LAPTOP" = true ] ; then
+	sudo systemctl enable tlp
+	sudo systemctl enable tlp-sleep
+	sudo systemctl mask systemd-rfkill.service
+	sudo systemctl mask systemd-rfkill.socket
+fi
+
